native_host/requirements.txt
pillow
numpy
requests
realesrgan
basicsr
pystray
torch
torchvision
torchaudio
